Allow CSharpSyntaxGenerator to be used as a source generator
This takes a fairly hacky approach to updating the code by converting
the CSharpSyntaxGenerator project to build twice: once as a
netcoreapp3.1 console app like it always did, and now also as a
netstandard2.0 library that is a source generator. The source generator
one only supports building the .cs outputs that are consumed as regular
APIs, and ignores the test or grammar outputs for now.

There is also no handling for cancellation yet which is not good for
performance.